fix exports
