+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/index.ts
@@ -1,3 +1,4 @@
+export * from './context';
 export * from './source-builder';
 export * from './stack-selector';
 export * from './types';
