+++ packages/@aws-cdk/toolkit-lib/lib/api/io/private/sdk-logger.ts
@@ -5,28 +5,6 @@ import { IO } from './messages';
 import { replacerBufferWithInfo } from '../../../private/util';
 import type { IoHelper } from '../../shared-private';
 
-/**
- * An SDK logging trace.
- *
- * Only info, warn and error level messages are emitted.
- * SDK traces are emitted as traces to the IoHost, but contain the original SDK logging level.
- */
-export interface SdkTrace {
-  /**
-   * The level the SDK has emitted the original message with
-   */
-  readonly sdkLevel: 'info' | 'warn' | 'error';
-
-  /**
-   * The content of the SDK trace
-   *
-   * This will include the request and response data for API calls, including potentially sensitive information.
-   *
-   * @see https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/logging-sdk-calls.html
-   */
-  readonly content: any;
-}
-
 export function asSdkLogger(ioHost: IoHelper): Logger {
   return new class implements Logger {
     // This is too much detail for our logs
