+++ packages/@aws-cdk/toolkit-lib/test/actions/synth.test.ts
@@ -1,6 +1,9 @@
 import { Toolkit } from '../../lib/toolkit';
 import { appFixture, builderFixture, TestIoHost } from '../_helpers';
 
+// these tests often run a bit longer than the default
+jest.setTimeout(10_000);
+
 const ioHost = new TestIoHost();
 const toolkit = new Toolkit({ ioHost });
